+++ src/apps/raycaster/renderer.rs
@@ -57,7 +57,7 @@ impl Renderer {
         }
     }
 
-    pub fn draw_sky(&self, dc: &mut DisplayController, _color: usize, _fog_color: usize) {
+    pub fn draw_sky(&self, dc: &mut DisplayController, _color: u8, _fog_color: u8) {
         dc.clear(DARK_GREY);
         dc.square(
             0,
@@ -112,7 +112,7 @@ impl Renderer {
         }
     }
 
-    pub fn draw_ground(&self, dc: &mut DisplayController, _color: usize, _fog_color: usize) {
+    pub fn draw_ground(&self, dc: &mut DisplayController, _color: u8, _fog_color: u8) {
         dc.square(
             0,
             (VIRTUAL_HEIGHT / 2) as isize,
@@ -167,8 +167,8 @@ impl Renderer {
         );
     }
 
-    pub fn scale_texture_column(&self, column: &Vec<usize>, new_height: isize) -> Vec<usize> {
-        let mut scaled: Vec<usize> = Vec::new();
+    pub fn scale_texture_column(&self, column: &Vec<u8>, new_height: isize) -> Vec<u8> {
+        let mut scaled: Vec<u8> = Vec::new();
 
         for pixel in 0..new_height as usize {
             let texture_pixel_index: usize = (pixel * column.len()) / new_height as usize;
@@ -179,7 +179,7 @@ impl Renderer {
         scaled
     }
 
-    pub fn shade_texture_column(&self, column: &mut [usize], map: &Map, dist: isize, x: isize) {
+    pub fn shade_texture_column(&self, column: &mut [u8], map: &Map, dist: isize, x: isize) {
         if dist > self.render_distance {
             column.fill(map.fog_color);
             return;
@@ -204,7 +204,7 @@ impl Renderer {
         }
     }
 
-    fn apply_level_1_shade(&self, column: &mut [usize], x: isize) {
+    fn apply_level_1_shade(&self, column: &mut [u8], x: isize) {
         for pixel in column.iter_mut().enumerate() {
             if (x % 2 == 0 && x % 4 == 0 && pixel.0 % 4 == 0)
                 || (x % 2 == 0 && (pixel.0 + 2) % 4 == 0)
@@ -214,7 +214,7 @@ impl Renderer {
         }
     }
 
-    fn apply_level_2_shade(&self, column: &mut [usize], x: isize) {
+    fn apply_level_2_shade(&self, column: &mut [u8], x: isize) {
         for pixel in column.iter_mut().enumerate() {
             if x % 2 == 0 && x % 4 == 0 && pixel.0 % 2 == 0 {
                 *pixel.1 = 0;
@@ -222,7 +222,7 @@ impl Renderer {
         }
     }
 
-    fn apply_level_3_shade(&self, column: &mut [usize], x: isize) {
+    fn apply_level_3_shade(&self, column: &mut [u8], x: isize) {
         for pixel in column.iter_mut().enumerate() {
             if (x % 2 == 0 && x % 4 == 0 && pixel.0 % 4 == 0)
                 || ((x + 2) % 2 == 0 && pixel.0 % 2 == 0)
@@ -232,7 +232,7 @@ impl Renderer {
         }
     }
 
-    fn apply_level_4_shade(&self, column: &mut [usize], x: isize) {
+    fn apply_level_4_shade(&self, column: &mut [u8], x: isize) {
         for pixel in column.iter_mut().enumerate() {
             if (x % 2 == 0 && pixel.0 % 2 == 0) || ((pixel.0 + 1) % 2 == 0) {
                 *pixel.1 = 0;
@@ -240,7 +240,7 @@ impl Renderer {
         }
     }
 
-    fn apply_level_5_shade(&self, column: &mut [usize], x: isize) {
+    fn apply_level_5_shade(&self, column: &mut [u8], x: isize) {
         for pixel in column.iter_mut().enumerate() {
             if (x % 2 == 0 && x % 4 == 0 && pixel.0 % 4 == 0)
                 || ((x + 1) % 2 == 0 && pixel.0 % 2 == 0)
@@ -250,7 +250,7 @@ impl Renderer {
         }
     }
 
-    fn apply_level_6_shade(&self, column: &mut [usize], x: isize) {
+    fn apply_level_6_shade(&self, column: &mut [u8], x: isize) {
         for pixel in column.iter_mut().enumerate() {
             if x % 2 == 0 && (x % 4 == 0 || pixel.0 % 2 == 0) {
                 *pixel.1 = 0;
@@ -258,7 +258,7 @@ impl Renderer {
         }
     }
 
-    fn apply_level_7_shade(&self, column: &mut [usize], x: isize) {
+    fn apply_level_7_shade(&self, column: &mut [u8], x: isize) {
         for pixel in column.iter_mut().enumerate() {
             if x % 2 == 0 && (x % 4 == 0 || pixel.0 % 4 != 0) {
                 *pixel.1 = 0;
@@ -270,8 +270,8 @@ impl Renderer {
         &self,
         dc: &mut DisplayController,
         x: isize,
-        data: &Vec<usize>,
-        transparent_color: usize,
+        data: &Vec<u8>,
+        transparent_color: u8,
     ) {
         let line_y_start = (VIRTUAL_HEIGHT / 2) as isize - data.len() as isize / 2;
         for pix_in_line in 0..data.len() as isize {
